// src/app/models/Customer.ts
export class Customer {
  fc_kdcust: string;
  fv_nama: string;
  fv_alamat: string;
  fv_email: string;
  fv_noktp: string;
  fc_jenis : string;
  fc_hold: string;
  fc_kota : string;
  fc_kdtop: string;
  fn_jumlah : string;
  fc_kdprofesi: string;
  fv_nmprofesi : string;
  fc_statusangsur : string;
}